DashboardStore: Moving the dashboard out of the mainStore. Drawer succesfully moved.
