+++ app/scripts/stores/mainStore.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { observable, computed, action, map, extendObservable } from 'mobx';
 import cookie from 'react-cookie';
 import authStore from '../stores/authStore';
+import dashboardStore from '../stores/dashboardStore';
 import provenanceStore from '../stores/provenanceStore';
 import transportLayer from '../transportLayer';
 import BaseUtils from '../util/baseUtils.js';
@@ -19,7 +20,7 @@ export class MainStore {
     @observable audit
     @observable currentUser
     @observable currentLocation
-    @observable drawer
+    // @observable drawer
     @observable destination
     @observable destinationKind
     @observable device
@@ -111,7 +112,7 @@ export class MainStore {
         this.counter = 0;
         this.currentLocation = null;
         this.currentUser = {};
-        this.drawer = observable.map();
+        // this.drawer = observable.map();
         this.device = {};
         this.destination = null;
         this.destinationKind = null;
@@ -428,40 +429,40 @@ export class MainStore {
             }
         }
     }
-    
-    @action toggleCollapseTree(router) {
-        if (this.drawer.get('collapsed')) {
-            this.downloadedItems.forEach((item) => {
-                item.open = true
-            })
-            this.drawer.set('collapsed', false);
-        } else {
-            this.listItems = this.projects
-            this.selectedItem = ''
-            this.downloadedItems.forEach((item) => {
-                item.open = false
-            })
-            router.push({pathname: ("/dashboard")})
-            this.drawer.set('collapsed', true);
-        }
-    }
-    
-    @action toggleDrawer() {
-        let drawerPosition = !this.drawer.get('open')
-        let contentStyle = { transition: 'margin-left 450ms cubic-bezier(0.23, 1, 0.32, 1)' };
-        drawerPosition ? contentStyle.marginLeft = this.drawer.get('width') : null
-        this.drawer.set('open', drawerPosition)
-        this.drawer.set('contentStyle', contentStyle)
-    }
-    
-    @action setDrawer() {
-        let width = 350;
-        let contentStyle = { transition: 'margin-left 450ms cubic-bezier(0.23, 1, 0.32, 1)' };
-        contentStyle.marginLeft = width;
-        this.drawer.set('open', true);
-        this.drawer.set('width', width);
-        this.drawer.set('contentStyle', contentStyle);
-    }
+    // 
+    // @action toggleCollapseTree(router) {
+    //     if (this.drawer.get('collapsed')) {
+    //         this.downloadedItems.forEach((item) => {
+    //             item.open = true
+    //         })
+    //         this.drawer.set('collapsed', false);
+    //     } else {
+    //         this.listItems = this.projects
+    //         this.selectedItem = ''
+    //         this.downloadedItems.forEach((item) => {
+    //             item.open = false
+    //         })
+    //         router.push({pathname: ("/dashboard")})
+    //         this.drawer.set('collapsed', true);
+    //     }
+    // }
+    // 
+    // @action toggleDrawer() {
+    //     let drawerPosition = !this.drawer.get('open')
+    //     let contentStyle = { transition: 'margin-left 450ms cubic-bezier(0.23, 1, 0.32, 1)' };
+    //     drawerPosition ? contentStyle.marginLeft = this.drawer.get('width') : null
+    //     this.drawer.set('open', drawerPosition)
+    //     this.drawer.set('contentStyle', contentStyle)
+    // }
+    // 
+    // @action setDrawer() {
+    //     let width = 350;
+    //     let contentStyle = { transition: 'margin-left 450ms cubic-bezier(0.23, 1, 0.32, 1)' };
+    //     contentStyle.marginLeft = width;
+    //     this.drawer.set('open', true);
+    //     this.drawer.set('width', width);
+    //     this.drawer.set('contentStyle', contentStyle);
+    // }
     
     @action toggleBackButtonVisibility(bool){
         this.showBackButton = bool;
